backend/requirements.txt
ffmpeg-python
google-genai
Pillow
faiss-cpu 
numpy
python-dotenv
fastapi
uvicorn[standard]
python-multipart
scenedetect[opencv]

